deploy/pyproject.toml
[project]
name = "tdx-prover-stack"
version = "0.1.0"
requires-python = ">=3.9,<4.0"
dependencies = [
    "aws-cdk-lib==2.186.0",
    "constructs>=10.0.0,<11.0.0",
    "python-dotenv>=1.0.1",
]
